--- a/ggml/src/ggml-qnn/npu/device/op_flash_attn.hpp
+++ b/ggml/src/ggml-qnn/npu/device/op_flash_attn.hpp
@@ -5,7 +5,9 @@
 namespace hexagon {
 
 bool flash_attn_f32(tensor * out, compute_params * params);
-bool is_flash_attn_supported(npu_device_tensor_op op, const npu_device_tensor_spec * dst,
-                             const npu_device_tensor_spec * srcs, size_t src_len);
+bool is_flash_attn_supported(npu_device_tensor_op           op,
+                             const npu_device_tensor_spec * dst,
+                             const npu_device_tensor_spec * srcs,
+                             size_t                         src_len);
 
 }  // namespace hexagon