## Overview
This PR implements performance optimizations for the Generalized Matrix-Vector Multiplication (GEMV) operations in llama.cpp. GEMV is a critical operation in the inference pipeline of large language models, especially during context processing and token generation.

## Changes
- Improved cache utilization through better memory access patterns
- Implemented vectorized operations using SIMD instructions where applicable
- Reduced unnecessary memory allocations and copies
- Optimized loop structures for better compiler auto-vectorization

## Performance Impact

Test setup: 8gen2
Test suite: `test-backend-ops`
Baseline: `6260c3166`
Optimized: `471ab1d87`

### Key Findings - GEMV Operations (n=1)

| Matrix Type | Baseline Performance | Optimized Performance | Improvement |
|-------------|---------------------|----------------------|------------|
| f32         | 8.23 GFLOPS         | 8.79 GFLOPS          | +6.8%      |
| f16         | 12.42 GFLOPS        | 12.99 GFLOPS         | +4.6%      |
| q4_0        | 16.33 GFLOPS        | 16.69 GFLOPS         | +2.2%      |
| q4_K        | 195.01 MFLOPS       | 195.65 MFLOPS        | +0.3%      |

### Summary

The GEMV optimization demonstrates meaningful performance improvements for the n=1 case, which is the specific target of this optimization:

1. **F32 format shows best improvement**: With a 6.8% speedup for matrix-vector operations, the optimization has the most significant impact on full-precision calculations

2. **F16 format also benefits**: The half-precision format shows a solid 4.6% performance improvement

3. **Quantized formats see modest gains**: The q4_0 format shows a 2.2% improvement, while q4_K sees minimal but still positive impact

This optimization is particularly beneficial for pure GEMV operations with n=1, which are critical for efficient token-by-token inference in large language models. The performance gains are consistent across different precision formats, with the most substantial improvements observed in full-precision calculations.

Log:

[test-backend-ops-perf-all.release.hexagon.471ab1d87.log](https://drive.google.com/file/d/1R5H8UGrNwHvfuXDcVAnpZ_CePzz3sShv/view?usp=sharing)
[test-backend-ops-perf-all.release.hexagon.6260c3166.log](https://drive.google.com/file/d/162NaKswLhxRdGAo0IicTfwQkV-4O7WEw/view?usp=sharing)

## Unit tests
Test setup: 8gen2
Test suite: `test-backend-ops`

```log
  8133/8133 tests passed
  Backend hexagon-npu: [1;32mOK[0m
Backend 2/2: CPU
  Skipping
2/2 backends passed
[1;32mOK[0m
```

Full log:
[test-backend-ops-all.release.hexagon.471ab1d87.7z](https://drive.google.com/file/d/1Xq0Rm6Yl5yGKjEzPBzs2zwaL-q5Y7XR8/view?usp=sharing)